requirements.txt
# requirements.txt الكامل (محدث للإصدارات الحديثة)
Flask==2.3.3
Flask-Bcrypt==1.0.1
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.3